+++ ASP/Test.aspx
@@ -1,4 +1,4 @@
-﻿<%@ Page Language="C#" %>
+﻿<%@ Page Language="C#" Title="Test" AutoEventWireup="true" CodeFile="~/Test.aspx.cs" Inherits="Test"%>
 
 <!DOCTYPE html>
 <html lang="en" xmlns="http://www.w3.org/1999/xhtml">
@@ -8,6 +8,7 @@
 </head>
 <body>
     <form id="form1" runat="server">   
+        <% getFestival(1)%>
     </form>
 </body>
 </html>
